+++ includes/utils/array_utils.h
@@ -19,7 +19,7 @@ T cumulative_sum(const std::vector<T> &counts, std::vector<T> &displacements, bo
 
   displacements.resize(counts.size());
   int sum = 0;
-  for (unsigned i{0}; i<counts.size(); ++i) {
+  for (size_t i{0}; i<counts.size(); ++i) {
     if (startAtZero)
       displacements[i] = sum;
     sum += counts[i];
@@ -49,7 +49,7 @@ std::vector<T> concatenate(const std::vector<T> &v1, const std::vector<T> &v2) {
 template <typename T>
 void split(const std::vector<T> &buffer, std::vector<std::vector<T>> &buffers, const std::vector<int> &displacements) {
   buffers.resize(displacements.size());
-  for (int p{0},i; p<displacements.size(); ++p) {
+  for (size_t p{0},i; p<displacements.size(); ++p) {
     int end_index = p<(displacements.size()-1) ? displacements[p+1] : buffer.size();
     buffers[p].resize(end_index-displacements[p]);
     for (i=0; i<buffers[p].size(); ++i)
